fix by tutors comments
