+++ src/components/app/app.jsx
@@ -2,7 +2,6 @@ import React, {PureComponent} from "react";
 import PropTypes from "prop-types";
 import {BrowserRouter, Route, Switch} from "react-router-dom";
 import {connect} from "react-redux";
-import {ActionCreator} from "../../reducer/reducer.js";
 
 import Main from "../main/main.jsx";
 import MoviePage from "../movie-page/movie-page.jsx";
@@ -26,7 +25,7 @@ class App extends PureComponent {
   }
 
   _renderMain() {
-    const {movieName, movieGenre, movieReleaseDate, movies, genresList, onFilterChange, currentGenre} = this.props;
+    const {movieName, movieGenre, movieReleaseDate, movies} = this.props;
 
     return (
       <Main
@@ -35,9 +34,6 @@ class App extends PureComponent {
         movieReleaseDate = {movieReleaseDate}
         onTitleClick = {this._titleClickHandler}
         movies = {movies}
-        genresList = {genresList}
-        onFilterChange = {onFilterChange}
-        currentGenre = {currentGenre}
       />
     );
   }
@@ -93,9 +89,6 @@ App.propTypes = {
         releaseYear: PropTypes.number.isRequired,
       }).isRequired
   ).isRequired,
-  currentGenre: PropTypes.string.isRequired,
-  onFilterChange: PropTypes.func.isRequired,
-  genresList: PropTypes.arrayOf(PropTypes.string.isRequired).isRequired,
   movie: PropTypes.oneOfType([
     PropTypes.number,
     PropTypes.shape({
@@ -124,17 +117,8 @@ App.propTypes = {
 };
 
 const mapStateToProps = (state) => ({
-  currentGenre: state.currentGenre,
-  genresList: state.genresList,
   movies: state.movies,
 });
 
-const mapDispatchToProps = (dispatch) => ({
-  onFilterChange(genre) {
-    dispatch(ActionCreator.getCurrentFilter(genre));
-    dispatch(ActionCreator.getFilteredMovies(genre));
-  }
-});
-
 export {App};
-export default connect(mapStateToProps, mapDispatchToProps)(App);
+export default connect(mapStateToProps)(App);
